Mascota asociada con cedula cliente

+++ demo/src/main/java/com/example/demo/repository/ClienteRepository.java
@@ -12,5 +12,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 public interface ClienteRepository extends JpaRepository<Cliente, Long>{
 
-    public Cliente findByCedula(Long cedula);
+    public Cliente findByCedula(int cedula);
 }
